[enhancement] support compress for large files
